refactor: CompanyDeliveryManagerAssignResult deliveryId 추가 및 result 패키지 이동
